Relocate polish tests to project test suite
