--- src/calibre/ebooks/oeb/polish/tests/__init__.py
@@ -1,5 +1,0 @@
-#!/usr/bin/env python
-
-
-__license__ = 'GPL v3'
-__copyright__ = '2013, Kovid Goyal <kovid at kovidgoyal.net>'
